refactor: symptom of disease for take data

+++ database/migrations/2024_12_12_223459_create_certainties_table.php
@@ -15,7 +15,7 @@ return new class extends Migration
             $table->id();
             $table->unsignedBigInteger('disease_id');
             $table->unsignedBigInteger('symptom_id');
-            $table->integer('value');
+            $table->decimal('value');
             $table->timestamps();
         });
     }
